[ref] directories and clear file name
